telegram_bot/requirements.txt
python-telegram-bot==20.7
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
alembic==1.13.0
openpyxl==3.1.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
